messaging_app/messaging_app/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
PyJWT==2.10.1
sqlparse==0.5.3
